Adicionado: autenticação e correção de bugs

+++ src/main/java/com/jolteam/financas/controller/DespesaController.java
@@ -3,6 +3,7 @@ package com.jolteam.financas.controller;
 import java.time.LocalDateTime;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -11,6 +12,7 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.jolteam.financas.dao.CategoriaDAO;
 import com.jolteam.financas.enums.TiposLogs;
@@ -19,9 +21,9 @@ import com.jolteam.financas.exceptions.DespesaException;
 import com.jolteam.financas.model.Categoria;
 import com.jolteam.financas.model.Despesa;
 import com.jolteam.financas.model.Log;
+import com.jolteam.financas.model.Usuario;
 import com.jolteam.financas.service.DespesaService;
 import com.jolteam.financas.service.LogService;
-import com.jolteam.financas.service.UsuarioService;
 
 @Controller
 public class DespesaController {
@@ -29,17 +31,19 @@ public class DespesaController {
 	@Autowired private DespesaService despesaService;
 	@Autowired private LogService logService;
 	@Autowired private CategoriaDAO categorias;
-	@Autowired private UsuarioService usuarioService;
 	
 	@GetMapping("/despesas/adicionar")
-	public ModelAndView viewDespesaAdicionar() {
+	public ModelAndView viewDespesaAdicionar(HttpSession session) {
 			ModelAndView mv=new ModelAndView("/despesas-adicionar");
 			mv.addObject("despesa", new Despesa());
-			mv.addObject("categorias", this.categorias.findByUsuarioAndTipoTransacao(this.usuarioService.getOne(1), TiposTransacoes.DESPESA));
+			mv.addObject("categorias", this.categorias.findByUsuarioAndTipoTransacaoOrderByDataCriacaoDesc((Usuario) session.getAttribute("usuarioLogado"), TiposTransacoes.DESPESA));
 			return mv;
 	}
+	
 	@PostMapping("/despesas/adicionar")
-	public ModelAndView adicionarDespesa(@ModelAttribute Despesa despesa,HttpServletRequest request) {
+	public ModelAndView adicionarDespesa(@ModelAttribute Despesa despesa,HttpServletRequest request, HttpSession session) {
+		despesa.setUsuario((Usuario) session.getAttribute("usuarioLogado"));
+		
 		try {
 			this.despesaService.salvar(despesa);
 			
@@ -49,22 +53,22 @@ public class DespesaController {
 			// salva um log de erro no banco
 			this.logService.save(new Log(despesa.getUsuario(), TiposLogs.ERRO_CADASTRO_DESPESA, LocalDateTime.now(), request.getRemoteAddr()));
 			
-			return this.viewDespesaAdicionar().addObject("erro", de.getMessage());
+			return this.viewDespesaAdicionar(session).addObject("erro", de.getMessage());
 		}
-		return this.viewDespesaAdicionar().addObject("sucesso", "Despesa salva com sucesso.");
+		return this.viewDespesaAdicionar(session).addObject("sucesso", "Despesa salva com sucesso.");
 	}
 	
-	
 	@GetMapping("/despesas/categorias")
-	public ModelAndView viewCategoriasDespesas() {
+	public ModelAndView viewCategoriasDespesas(HttpSession session) {
 		ModelAndView mv= new ModelAndView("despesas-categorias");
 		mv.addObject("catDespesa", new Categoria());
-		mv.addObject("listacatDespesa", this.categorias.findAllByTipoTransacao(TiposTransacoes.DESPESA));
+		mv.addObject("listaCatDespesa", this.categorias.findByUsuarioAndTipoTransacaoOrderByDataCriacaoDesc((Usuario) session.getAttribute("usuarioLogado"), TiposTransacoes.DESPESA));
 		return mv;
 	}
 	
-	@PostMapping("/despesas/categorias/adicionar")
-	public ModelAndView adicionarCatDespesa(@ModelAttribute Categoria catDespesa,HttpServletRequest request) {
+	@PostMapping("/despesas/categorias")
+	public ModelAndView adicionarCatDespesa(@ModelAttribute Categoria catDespesa,HttpServletRequest request, HttpSession session) {
+		catDespesa.setUsuario((Usuario) session.getAttribute("usuarioLogado"));
 		
 		try {
 			this.despesaService.salvarCategoriaDespesa(catDespesa);
@@ -73,14 +77,26 @@ public class DespesaController {
 		} catch (DespesaException de) {
 			this.logService.save(new Log(catDespesa.getUsuario(), TiposLogs.ERRO_CADASTRO_CATEGORIA_DESPESA, LocalDateTime.now(), request.getRemoteAddr()));
 			
-			return this.viewCategoriasDespesas().addObject("erro", de.getMessage());
+			return this.viewCategoriasDespesas(session).addObject("erro", de.getMessage());
 
 		}
-		return this.viewCategoriasDespesas().addObject("sucesso", "Categoria salva com sucesso.");
+		return this.viewCategoriasDespesas(session).addObject("sucesso", "Categoria salva com sucesso.");
 	}
+	
 	@GetMapping("/despesas/categorias/excluir")
-	public String excluirCatDespesa(@RequestParam Integer id) {
-		this.categorias.deleteById(id);
+	public String excluirCatDespesa(@RequestParam Integer id, RedirectAttributes ra) {
+		try {
+			Categoria catExistente = this.categorias.findById(id).orElseThrow(() -> new Exception("Categoria inexistente."));
+			
+			if (this.despesaService.existsByCategoria(catExistente)) {
+				ra.addFlashAttribute("msgErro", "A categoria selecionada contém vínculos com outras informações.");
+			} else {
+				this.categorias.deleteById(id);
+			}
+		} catch (Exception e) {
+			System.out.println(e.getMessage());
+		}
+		
 		return "redirect:/despesas/categorias";
 	}
